Remove cairo-hint file
