+++ src/tests/mod.rs
@@ -31,8 +31,6 @@ mod cairo_run_test;
 mod pedersen_test;
 mod struct_test;
 
-mod cairo_1_hints;
-
 #[cfg(feature = "skip_next_instruction_hint")]
 mod skip_instruction_test;
